[Bronze V] Title: 이상한 기호, Time: 76 ms, Memory: 5520 KB -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -1,17 +1,17 @@
 class Program
 {
-    static int result;
+    static long result;
     static void Main(string[] args)
     {
         string s = Console.ReadLine();
         string[]ss = s.Split();
-        int a = Convert.ToInt32(ss[0]);
-        int b = Convert.ToInt32(ss[1]);
+        long a = Convert.ToInt32(ss[0]);
+        long b = Convert.ToInt32(ss[1]);
         result = Calcul(a, b);
         Console.WriteLine(result);
     }
 
-    static int Calcul(int a, int b)
+    static long Calcul(long a, long b)
     {
         return result = (a + b) * (a - b);
     }
